chore(apps): apply review conventions to the replay-vision fixtures
