--- apps/replay-vision/fastapi/inventory-api/requirements.txt
@@ -1,2 +1,0 @@
-fastapi>=0.115
-uvicorn>=0.34
