+++ apps/replay-vision/javascript-node/inventory-api/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "inventory-api",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "start": "node src/server.mjs"
+  }
+}
