+++ apps/replay-vision/next-js/artisan-ceramics/.env.example
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_POSTHOG_PROJECT_TOKEN=
+NEXT_PUBLIC_POSTHOG_HOST=https://us.i.posthog.com
