+++ apps/replay-vision/next-js/artisan-ceramics/lib/products.ts
@@ -1,0 +1,5 @@
+export const PRODUCTS = [
+  { slug: 'speckled-mug', name: 'Speckled mug', priceUsd: 28 },
+  { slug: 'serving-bowl', name: 'Serving bowl', priceUsd: 54 },
+  { slug: 'bud-vase', name: 'Bud vase', priceUsd: 36 },
+]
