--- apps/replay-vision/next-js/artisan-checkout/app/checkout/success/page.tsx
@@ -1,11 +1,0 @@
-import Link from 'next/link'
-
-export default function CheckoutSuccessPage() {
-  return (
-    <main>
-      <h1>Order placed</h1>
-      <p>Thanks for supporting the studio.</p>
-      <Link href="/">Back to the shop</Link>
-    </main>
-  )
-}
