--- apps/replay-vision/next-js/artisan-checkout/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "artisan-checkout",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "next": "^15.4.0",
-    "posthog-js": "^1.250.0",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0"
-  },
-  "devDependencies": {
-    "@types/node": "^22.15.0",
-    "@types/react": "^19.1.0",
-    "typescript": "^5.8.0"
-  }
-}
