+++ apps/replay-vision/react-vite/court-booking/index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Court Booking</title>
+  </head>
+  <body style="font-family: system-ui; max-width: 640px; margin: 2rem auto">
+    <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
+  </body>
+</html>
