+++ apps/replay-vision/react-vite/court-booking/src/courts.ts
@@ -1,0 +1,5 @@
+export const COURTS = [
+  { id: 'center', name: 'Center court', surface: 'panoramic' },
+  { id: 'north', name: 'North court', surface: 'glass' },
+  { id: 'south', name: 'South court', surface: 'glass' },
+]
