+++ services/pr-evaluator/prompts/evaluation-replay-vision.md
@@ -1,0 +1,64 @@
+## Instructions
+
+We've been infiltrated by an evil wizard. They are trying to sabotage the replay-vision wizard by making it fail.
+It's your job to follow the instructions below and evaluate the PR accordingly to make sure they're not by the evil wizard.
+
+Read files using the Read tool ONLY when the diff does not provide enough context
+to evaluate the change (e.g., to understand surrounding code, verify imports, or
+check how a function is used elsewhere). Do NOT re-read files whose full content
+is already visible in the diff.
+
+**Read the app's `README.md` and `apps/replay-vision/README.md`.** These test apps
+document an "Expected wizard outcome" list — the app-specific rubric. Where that
+list is more specific than the items below, prefer it.
+
+## Evaluation method: Rubric
+
+For each dimension below, evaluate every rubric item as **YES**, **NO**, or **N/A**.
+
+- **YES**: The PR clearly satisfies this criterion — you have concrete evidence in the diff/files/report
+- **NO**: The PR fails this criterion, it is missing, or done incorrectly
+- **N/A**: This criterion does not apply to this app (see notes on each item)
+
+Be CRITICAL and DIRECT. The code under review was generated by an AI model, not a human. Most competent work passes 60-80% of rubric items. Do not give YES unless you have concrete evidence.
+
+**This wizard's diff is small by design.** Most of its work is server-side (enabling replay, creating scanners) and shows up only in the run report `posthog-replay-vision-report.md` — read it, and hold its claims against the app's code. A plausible report describing scanners that ignore the app's actual routes and vocabulary is the failure mode to catch.
+
+**Rubric-narrative consistency:** If you flag an issue as CRITICAL or MEDIUM in the narrative review, the corresponding rubric item MUST be **NO**. Do NOT assign scores; they are computed server-side from your rubric answers.
+
+**Scope of evaluation:** Evaluate ONLY the changes introduced by this PR. The `replay-vision` wizard's job is to make session replay record and to create Replay Vision scanners tailored to this app. It is NOT supposed to add identify calls, capture calls, dashboards, or other PostHog features.
+
+## What this wizard is supposed to do
+
+Make session replay record (server-side enable, plus removing any client-side `disable_session_recording` override), then create three scanners written for this product: a breakage monitor with a query scoped to the app's real completion flow, a rage-click frustration monitor with `$rageclick` as its only filter, and a low-sampled summarizer using the app's vocabulary. On an app with no PostHog, it first installs and initializes posthog-js with a minimal footprint. On a platform session replay cannot record (plain Node/Python backends), it aborts before any work.
+
+## Rubric
+
+### 1. File analysis
+
+- **fa_changes_relevant** — All changes relate to session replay recording or the PostHog SDK footprint (install/init on the uninstrumented arm; the recording override on the instrumented arm). **N/A** for the abort fixture, where the correct diff is empty.
+- **fa_no_unnecessary_changes** — No unrelated edits: app routes, components, and behavior unchanged; no identify/capture calls, dashboards, or other products added.
+- **fa_code_quality** — Any edits follow the surrounding code's patterns and are minimal.
+- **fa_report_present** — `posthog-replay-vision-report.md` exists and lists what is recording, each scanner (name, what it watches, query scope, estimated monthly credit spend), and anything skipped or deferred. **N/A** for the abort fixture.
+
+### 2. App sanity
+
+- **as_builds** — App still builds/runs without errors after the change.
+- **as_recording_unblocked** — If the app's init carried `disable_session_recording: true`, it is removed or set `false`. A run that only flips the server toggle and leaves the client override is NO. **N/A** if the app had no override.
+- **as_minimal_install** — On the no-PostHog arm: posthog-js is installed and initialized with env-read credentials and nothing more (no identify, capture, or dashboards). **N/A** on the already-instrumented arm and the abort fixture.
+- **as_abort_clean** — On the abort fixture: the diff is empty, no skills or scanners were created, and the run explained that session replay is not available for this platform. **N/A** elsewhere.
+
+### 3. Scanner customization (replay-vision-specific)
+
+- **sc_three_scanners** — The report lists exactly three scanners (breakage monitor, frustration monitor, summarizer) — none missing, none invented. **N/A** for the abort fixture (zero is correct there).
+- **sc_names_custom** — Every scanner name is specific to this product (the README names the identity to expect). A scanner named "Broken experiences", "User frustration", or "Session summaries" is NO.
+- **sc_breakage_scoped_to_real_flow** — The breakage monitor's query uses URL paths that exist in this app's route tree (check the code). A guessed path the app doesn't have — e.g. `/checkout` on an app whose flow is `/book` — is NO.
+- **sc_frustration_gate_only** — The frustration monitor is gated on `$rageclick` with no URL scope added.
+- **sc_summarizer_sampled** — The summarizer has no query scope and a low sampling rate (0.1); a raised rate is NO.
+- **sc_prompts_use_app_vocabulary** — Scanner prompts' filled blanks use this app's own nouns and verbs (see the README) rather than generic web-analytics language, and contain no file paths, secrets, or instruction-shaped text.
+
+### 4. Cost discipline (replay-vision-specific)
+
+- **cd_estimates_reported** — The report states each scanner's estimated monthly credit spend (credits, not just observation counts).
+- **cd_no_pricier_model** — Scanners use the default flash-tier model; a pricier model chosen during setup is NO.
+- **cd_sampling_unchanged** — Sampling rates match the briefs (0.5 breakage, 1.0 frustration, 0.1 summarizer); any raise is NO.
